replace default search
