+++ pantheon-bundle/frontend/src/app/search.tsx
@@ -1,362 +1,581 @@
-import React, { Component, FormEvent } from "react"
+import React, { Component } from "react";
 import {
-  Alert, AlertActionCloseButton,
-  DataList, DataListItem, DataListItemRow, DataListItemCells,
-  DataListCell, Button, Modal,
-  Level, LevelItem, Checkbox, ModalVariant
-} from "@patternfly/react-core"
-import "@app/app.css"
+  Drawer,
+  DrawerPanelContent,
+  DrawerContent,
+  DrawerContentBody,
+  DrawerPanelBody,
+  DrawerHead,
+  DrawerActions,
+  DrawerCloseButton,
+  Button, ButtonVariant,
+  InputGroup,
+  Select,
+  SelectOption,
+  Dropdown,
+  DropdownItem,
+  DropdownSeparator,
+  KebabToggle,
+  Toolbar, ToolbarItem, ToolbarContent, ToolbarFilter, ToolbarToggleGroup, ToolbarGroup,
+  TextInput,
+  SelectVariant,
+  ExpandableSection,
+  Checkbox,
+  Divider,
+  SimpleListItem,
+  SimpleList,
+  SearchInput,
+  Alert,
+  ToolbarChipGroup,
+  ToolbarChip,
+
+} from "@patternfly/react-core";
+
+import { SearchResults } from "@app/searchResults";
 import { BuildInfo } from "./components/Chrome/Header/BuildInfo"
-import { Pagination } from "@app/Pagination"
-import { BrowserRouter as Router, Route, Link, Switch } from "react-router-dom"
+
+import "@app/app.css";
+import SearchIcon from "@patternfly/react-icons/dist/js/icons/search-icon";
+import FilterIcon from "@patternfly/react-icons/dist/js/icons/filter-icon";
 import { IAppState } from "@app/app"
-import { SearchFilter } from "@app/searchFilter";
-import SpinImage from "@app/images/spin.gif";
-import { Fields, SlingTypesPrefixes } from "@app/Constants";
 
 export interface ISearchState {
-  alertOneVisible: boolean
-  checkNextPageRow: string
-  columns: string[]
-  contentType: string
-  filterQuery: string
-  isEmptyResults: boolean
-  isModalOpen: boolean
-  isSearchException: boolean
-  displayLoadIcon: boolean
-  moduleName: string
-  modulePath: string
-  moduleType: string
-  moduleUpdatedDate: string
-  nextPageRowCount: number
-  page: number
-  pageLimit: number
-  redirect: boolean
-  redirectLocation: string
-  results: any
-  selectAllCheckValue: boolean
-  showDropdownOptions: boolean
-  sortKey: string
-}
+  filterLabel: string
+  isExpanded: boolean
+  assembliesIsExpanded: boolean
+  expandableSectionIsExpanded: boolean
+  modulesIsExpanded: boolean
+  productFilterIsExpanded: boolean
+  repoFilterIsExpanded: boolean
+  products: Array<{ name: string, id: string }>
+  repositories: Array<{ name: string, id: string, checked: boolean }>
+  filteredRepositories: Array<{ name: string, id: string, checked: boolean }>
 
+  inputValue: string,
+  statusIsExpanded: boolean,
+  ctypeIsExpanded: boolean,
+  filters: {
+    ctype: any,
+    status: any
+  },
+
+  productFilterValue: string
+  repoFilterValue: string
+
+  productsSelected: string[]
+  repositoriesSelected: string[]
+}
 class Search extends Component<IAppState, ISearchState> {
-  public static KEY_CHECKEDITEM: string = "checkedItem"
-  public static KEY_TRANSIENTPATH: string = "pant:transientPath"
-  public published = "-"
+  private drawerRef: React.RefObject<HTMLInputElement>;
 
   constructor(props) {
     super(props);
     this.state = {
-      alertOneVisible: true,
-      checkNextPageRow: "",
-      columns: ["Name", "Description", "Source Type", "Source Name", "Upload Time"],
-      contentType: "module",
-      displayLoadIcon: true,
-      filterQuery: "",
-      isEmptyResults: false,
-      isModalOpen: false,
-      isSearchException: false,
-      moduleName: "",
-      modulePath: "",
-      moduleType: "",
-      moduleUpdatedDate: "",
-      nextPageRowCount: 1,
-      page: 1,
-      pageLimit: 25,
-      redirect: false,
-      redirectLocation: "",
-      results: [{ "pant:transientPath": "", "pant:dateUploaded": "", "name": "", "jcr:title": "", "jcr:description": "", "sling:transientSource": "", "pant:transientSourceName": "", "checkedItem": false,"publishedDate": "-","pant:moduleType": "-", "variant": ""}],
-      selectAllCheckValue: false,
-      showDropdownOptions: true,
-      sortKey: ""
+      // states for drawer
+      filterLabel: "repo",
+      isExpanded: true,
+      assembliesIsExpanded: true,
+      expandableSectionIsExpanded: true,
+      modulesIsExpanded: true,
+      productFilterIsExpanded: true,
+      repoFilterIsExpanded: true,
+      products: [{ name: "", id: "" }],
+      repositories: [{ name: "", id: "", checked: false }],
+      filteredRepositories: [{ name: "", id: "", checked: false }],
+      // states for toolbar
+      inputValue: "",
+      statusIsExpanded: false,
+      ctypeIsExpanded: false,
+      filters: {
+        ctype: [],
+        status: []
+      },
+
+      // filters
+      productFilterValue: "",
+      repoFilterValue: "",
+
+      // search
+      productsSelected: [],
+      repositoriesSelected: [],
     };
+    this.drawerRef = React.createRef();
+
   }
 
   public componentDidMount() {
-    this.doSearch()
+
+    // list repos inside the drawer
+    this.getRepositories()
+    // this.getProducts()
+
+    // TODO: enable resize
+    // toolbar
+    // window.addEventListener("resize", this.closeExpandableContent);
   }
 
+  public componentWillMount() {
+    // list repos inside the drawer
+    this.getRepositories()
+    // this.getProducts()
+  }
+
+  public componentWillUnmount() {
+    // TODO: enable resize
+    // toolbar
+    // window.removeEventListener("resize", this.closeExpandableContent);
+  }
   public render() {
-    const { isEmptyResults } = this.state;
+    const { filterLabel, isExpanded, assembliesIsExpanded, modulesIsExpanded, productFilterIsExpanded, repoFilterIsExpanded, expandableSectionIsExpanded, repositories, inputValue, filters, statusIsExpanded, ctypeIsExpanded } = this.state;
+
+    const panelContent = (
+      <DrawerPanelContent widths={{ lg: "width_25" }}>
+        <DrawerHead>
+          <span className="pf-c-title pf-m-2xl" tabIndex={isExpanded ? 0 : -1} ref={this.drawerRef}>Filters</span>
+          <DrawerActions>
+            <DrawerCloseButton onClick={this.onCloseClick} />
+          </DrawerActions>
+          <ExpandableSection className="filters-drawer filters-drawer--by-repo" toggleText="By repository" isActive={true} isExpanded={repoFilterIsExpanded} onToggle={this.onRepositoriesToggle}>
+            <SearchInput
+              placeholder="Filter"
+              value={this.state.repoFilterValue}
+              onChange={this.onChangeRepoFilter}
+              onClear={(evt) => this.onChangeRepoFilter("", evt)}
+              className='filters-drawer__repo-search'
+            />
+            {this.state.filteredRepositories && this.state.filteredRepositories.length > 0 &&
+              <SimpleList aria-label="Repository List">
+                {this.state.filteredRepositories.map((data) => (
+                  <SimpleListItem key={data.id} className='repo-list filters-drawer__repo-list'>
+                    <Checkbox label={data.name} aria-label="uncontrolled checkbox" id={data.id} onChange={this.onSelectRepositories} isChecked={data.checked}/>
+                  </SimpleListItem>
+                ))}
+              </SimpleList>
+            }
+
+          </ExpandableSection>
+          <br />
+          {/* <ExpandableSection toggleText="By product">
+            <SearchInput
+              placeholder="Filter"
+              value={this.state.productFilterValue}
+              onChange={this.onChangeProductFilter}
+              onClear={(evt) => this.onChangeProductFilter("", evt)}
+            />
+            <SimpleList aria-label="Product List">
+              {this.state.products.map((data) => (
+                <SimpleListItem key={data.id}>
+                  <Checkbox label={data.name} aria-label="uncontrolled checkbox" id={data.id} />
+                </SimpleListItem>
+              ))}
+            </SimpleList>
+
+          </ExpandableSection> */}
+        </DrawerHead>
+      </DrawerPanelContent>
+    );
+    const drawerContent = (
+      <React.Fragment>
+        <ExpandableSection toggleText="Modules" className="pf-c-title search-results__section search-results__section--module" isActive={true} isExpanded={modulesIsExpanded} onToggle={this.onModulesToggle}>
+          <SearchResults
+            contentType="module"
+            keyWord={this.state.inputValue}
+            repositoriesSelected={this.state.repositoriesSelected}
+            productsSelected={this.state.productsSelected}
+            userAuthenticated={this.props.userAuthenticated}
+            filters={this.state.filters}
+          />
+
+        </ExpandableSection>
+        <br />
+        <ExpandableSection toggleText="Assemblies" className="pf-c-title search-results__section search-results__section--assembly" isActive={true} isExpanded={assembliesIsExpanded} onToggle={this.onAssembliesToggle}>
+          <SearchResults
+            contentType="assembly"
+            keyWord={this.state.inputValue}
+            repositoriesSelected={this.state.repositoriesSelected}
+            productsSelected={this.state.productsSelected}
+            userAuthenticated={this.props.userAuthenticated}
+            filters={this.state.filters}
+          />
+
+        </ExpandableSection>
+      </React.Fragment>
+    );
+
+    const statusMenuItems = [
+      <SelectOption key="statusDraft" value="draft" label="Draft" className="dropdown-filter__option dropdown-filter__option--status dropdown-filter__option--draft" />,
+      <SelectOption key="statusPublished" value="published" label="Published" className="dropdown-filter__option dropdown-filter__option--status dropdown-filter__option--published" />
+    ];
+
+    const contentTypeMenuItems = [
+      <SelectOption key="ctypeConcept" value="CONCEPT" label="Concept" className="dropdown-filter__option dropdown-filter__option--content-type dropdown-filter__option--concept" />,
+      <SelectOption key="ctypeProcedure" value="PROCEDURE" label="Procedure" className="dropdown-filter__option dropdown-filter__option--content-type dropdown-filter__option--procedure" />,
+      <SelectOption key="ctypeReference" value="REFERENCE" label="Reference" className="dropdown-filter__option dropdown-filter__option--content-type dropdown-filter__option--reference" />
+    ];
+
+    const toggleGroupItems = (
+      <React.Fragment>
+        <ToolbarItem id="filters-bar__toolbar-toggle">
+          <Button variant="tertiary" aria-expanded={isExpanded} onClick={this.onClick} icon={<FilterIcon />} />
+        </ToolbarItem>
+        <ToolbarItem>
+          <InputGroup>
+            <SearchInput
+              className="filters-bar__name-search"
+              name="textInput"
+              id="textInput"
+              placeholder="Find by name"
+              type="search"
+              aria-label="search input"
+              onChange={this.onInputChange}
+              onClear={this.onInputClear}
+              value={inputValue}
+            />
+            <Button variant={ButtonVariant.control} aria-label="search button for search input">
+              <SearchIcon />
+            </Button>
+          </InputGroup>
+        </ToolbarItem>
+        <ToolbarGroup variant="filter-group">
+          <ToolbarFilter
+            chips={filters.status}
+            deleteChip={this.onDelete}
+            deleteChipGroup={this.onDeleteGroup}
+            categoryName="Status"
+            className="dropdown-filter filters-bar__dropdown-filter filters-bar__dropdown-filter--status"
+          >
+            <Select
+              variant={SelectVariant.checkbox}
+              aria-label="Status"
+              onToggle={this.onStatusToggle}
+              onSelect={this.onStatusSelect}
+              selections={filters.status}
+              isOpen={statusIsExpanded}
+              placeholderText="Status"
+            >
+              {statusMenuItems}
+            </Select>
+          </ToolbarFilter>
+          <ToolbarFilter chips={filters.ctype} deleteChipGroup={this.onDeleteGroup} deleteChip={this.onDelete} categoryName="Content Type" className="dropdown-filter filters-bar__dropdown-filter filters-bar__dropdown-filter--content-type">
+            <Select
+              variant={SelectVariant.checkbox}
+              aria-label="Content Type"
+              onToggle={this.onCtypeToggle}
+              onSelect={this.onCtypeSelect}
+              selections={filters.ctype}
+              isOpen={ctypeIsExpanded}
+              placeholderText="Content Type"
+            >
+              {contentTypeMenuItems}
+            </Select>
+          </ToolbarFilter>
+        </ToolbarGroup>
+      </React.Fragment>
+    );
+
+    // const dropdownItems = [
+    //   <DropdownItem key="link">Link</DropdownItem>,
+    //   <DropdownItem key="action" component="button">
+    //     Action
+    //       </DropdownItem>,
+    //   <DropdownItem key="disabled link" isDisabled={true}>
+    //     Disabled Link
+    //       </DropdownItem>,
+    //   <DropdownItem key="disabled action" isDisabled={true} component="button">
+    //     Disabled Action
+    //       </DropdownItem>,
+    //   <DropdownSeparator key="separator" />,
+    //   <DropdownItem key="separated link">Separated Link</DropdownItem>,
+    //   <DropdownItem key="separated action" component="button">
+    //     Separated Action
+    //       </DropdownItem>
+    // ];
+
+    const toolbarItems = (
+      <React.Fragment>
+        <ToolbarToggleGroup toggleIcon={<FilterIcon />} breakpoint="xl">
+          {toggleGroupItems}
+        </ToolbarToggleGroup>
+        <ToolbarGroup variant="icon-button-group">
+        </ToolbarGroup>
+        {/* <ToolbarItem>
+          <Dropdown
+            toggle={<KebabToggle onToggle={this.onKebabToggle} />}
+            isOpen={kebabIsOpen}
+            isPlain={true}
+            dropdownItems={dropdownItems}
+          />
+        </ToolbarItem> */}
+      </React.Fragment>
+    );
 
     return (
       <React.Fragment>
-        <SearchFilter
-          onKeyDown={this.getRows}
-          onClick={this.newSearch}
-          filterQuery={this.setQuery}
-          // we could add one that monitors for changes and when is tru we run getRows. To discuss.
-        />
-        <div className="notification-container">
-          <Pagination
-            handleMoveLeft={this.updatePageCounter("L")}
-            handleMoveRight={this.updatePageCounter("R")}
-            handleMoveToFirst={this.updatePageCounter("F")}
-            pageNumber={this.state.page}
-            nextPageRecordCount={this.state.nextPageRowCount}
-            handlePerPageLimit={this.changePerPageLimit}
-            perPageLimit={this.state.pageLimit}
-            showDropdownOptions={this.state.showDropdownOptions}
-            bottom={false}
-            className="pagination"
-          />
-        </div>
-        {/* @todo This should _definitely_ be a table */}
-        <DataList aria-label="Search results Headings" >
-          <DataListItem aria-labelledby="span-name">
-            <DataListItemRow id="data-rows-header" >
-              {/* {this.props.userAuthenticated && !this.state.isEmptyResults &&
-                <Checkbox aria-labelledby="width-ex1-check1"
-                  className="checkbox"
-                  isChecked={this.state.selectAllCheckValue}
-                  checked={this.state.selectAllCheckValue}
-                  aria-label="controlled checkbox example"
-                  id="check"
-                  onChange={this.handleSelectAll}
-                />} */}
-              <DataListItemCells
-                dataListCells={[
-                  <DataListCell width={2} key="title">
-                    <span className="sp-prop-nosort" id="span-name" aria-label="column name">
-                      Title
-                    </span>
-                  </DataListCell>,
-                  <DataListCell key="resource source">
-                    <span className="sp-prop-nosort" id="span-source-type">
-                      Published
-                    </span>
-                  </DataListCell>,
-                  <DataListCell key="source name">
-                    <span className="sp-prop-nosort" id="span-source-name">
-                      Draft Uploaded
-                    </span>
-                  </DataListCell>,
-                  <DataListCell key="upload time">
-                    <span className="sp-prop-nosort" id="span-upload-time" aria-label="column upload time">
-                      Content Type
-                    </span>
-                  </DataListCell>,
-                ]}
-              />
-              </DataListItemRow>
-              {this.state.displayLoadIcon && (
-                <Level hasGutter={true}>
-                  <LevelItem />
-                  <LevelItem>
-                      <div className="notification-container">
-                        {/* @todo Take out these <br>s and replace it with CSS layout */}
-                        <br />
-                        <br />
-                          <img src={SpinImage} alt="Spinlogo"/>
-                        <br />
-                        <br />
-                      </div>
-                    </LevelItem>
-                    <LevelItem />
-                  </Level>
-                )}                          
-                {!this.state.displayLoadIcon && (this.state.results.map((data, key) => (
-                  <DataListItemRow id="data-rows" key={key}>
-                    <DataListItemCells key={"cells_" + key}
-                      dataListCells={[
-                        <DataListCell key={"title_" + key} width={2}>
-                          {this.props.userAuthenticated &&
-                            <Link to={data["sling:resourceType"].substring(SlingTypesPrefixes.PANTHEON.length) + "/" + data["pant:transientPath"] + "?variant=" + data.variant} key={"link_" + key}>
-                              {data["jcr:title"] !== "-" ? data["jcr:title"] : data["pant:transientPath"]}
-                            </Link>
-                          }
-                          {!this.props.userAuthenticated &&
-                            <a href={"/pantheon/preview/released/" + data["jcr:uuid"]} target="_blank">
-                              {data["jcr:title"] !== "-" ? data["jcr:title"] : data["pant:transientPath"]}
-                            </a>
-                          }
-                        </DataListCell>,      
-                        <DataListCell key={"published-date_" + key}>                          
-                          <span>{data[Fields.PANT_PUBLISHED_DATE]}</span>
-                        </DataListCell>,
-                        <DataListCell key={"date-uploaded_" + key}>                              
-                          <span>{data[Fields.PANT_DATE_UPLOADED]}</span>
-                        </DataListCell>,
-                        <DataListCell key={"module-type_" + key}>
-                          <span >{data[Fields.PANT_MODULE_TYPE]}</span>
-                        </DataListCell>
-                      ]}
-                    />
-                  </DataListItemRow>
-                )))}
-
-                {isEmptyResults && (
-                  <Level hasGutter={true}>
-                    <LevelItem />
-                    <LevelItem>
-                      <div className="notification-container">
-                        <br />
-                        <br />
-                        <Alert
-                          variant="warning"
-                          title={"No modules found with your search"}
-                          actionClose={<AlertActionCloseButton onClose={this.dismissNotification} />}
-                        />
-                        <br />
-                        <br />
-                      </div></LevelItem>
-                    <LevelItem />
-                  </Level>
-
-                )}
-                {this.state.isSearchException && (
-                  <Level hasGutter={true}>
-                    <LevelItem />
-                    <LevelItem>
-                      <div className="notification-container">
-                        <br />
-                        <br />
-                        <Alert
-                          variant="danger"
-                          title={"Error in fetching search results"}
-                          actionClose={<AlertActionCloseButton onClose={this.dismissNotification} />}
-                        />
-                        <br />
-                        <br />
-                      </div>
-                      </LevelItem>
-                    <LevelItem />
-                  </Level>
-                )}
-              </DataListItem>
-            </DataList>
-            <div className="notification-container">
-              <Pagination
-                handleMoveLeft={this.updatePageCounter("L")}
-                handleMoveRight={this.updatePageCounter("R")}
-                handleMoveToFirst={this.updatePageCounter("F")}
-                pageNumber={this.state.page}
-                nextPageRecordCount={this.state.nextPageRowCount}
-                handlePerPageLimit={this.changePerPageLimit}
-                perPageLimit={this.state.pageLimit}
-                showDropdownOptions={!this.state.showDropdownOptions}
-                bottom={true}
-                className="pagination"
-              />
-              <BuildInfo />
-            </div>
+        <Alert variant="info" title="Beta feature" >
+          <p>
+            Please give us your feedback {"  "}
+            <a href="https://projects.engineering.redhat.com/browse/CCS-3969" target="_blank">here.</a>
+          </p>
+        </Alert>
+        <br />
+        <Toolbar
+          id="toolbar-with-filter"
+          className="pf-m-toggle-group-container filters-bar__filters-wrapper"
+          collapseListedFiltersBreakpoint="xl"
+          clearAllFilters={this.onDelete}
+        >
+          <ToolbarContent>{toolbarItems}</ToolbarContent>
+        </Toolbar>
+        <Divider />
+        <Drawer isExpanded={isExpanded} isInline={true} position="left" onExpand={this.onExpand}>
+          <DrawerContent panelContent={panelContent} width="width_50">
+            <DrawerContentBody className="search-results" width="width_50">
+              {drawerContent}
+            </DrawerContentBody>
+          </DrawerContent>
+        </Drawer>
+        <BuildInfo />
       </React.Fragment>
     );
   }
 
-  private handleSelectAll = (checked: boolean, event: FormEvent<HTMLInputElement>) => {
-    const newResults: any[] = []
-    this.state.results.map(dataitem => {
-      newResults.push(JSON.parse(JSON.stringify(dataitem))) // clones the object
-      newResults[newResults.length - 1][Search.KEY_CHECKEDITEM] = checked
-    })
-
-    this.setState({
-      results: newResults,
-      selectAllCheckValue: checked
-    })
-  }
-
-  private getRows = (event) => {
-    if (event.key === "Enter") {
-      this.newSearch()
-    }
-  };
-
-  private newSearch = () => {
-    this.setState({ page: 1 }, () => {
-      this.doSearch()
-    })
-  }
-
-  // Handle gateway timeout on slow connections.
-  private doSearch = () => {
-    this.setState({ displayLoadIcon: true })
-    fetch(this.buildSearchUrl())
-      .then(response => response.json())
-      .then(responseJSON => {
-        this.setState({ results: responseJSON.results, nextPageRowCount: responseJSON.hasNextPage ? 1 : 0 })
-      })
-      .then(() => {
-        if (JSON.stringify(this.state.results) === "[]") {
-          this.setState({
-            displayLoadIcon: false,
-            isEmptyResults: true,
-            selectAllCheckValue: false
-          })
+  // methods for drawer
+  private getRepositories = () => {
+    const path = "/content/repositories.harray.1.json"
+    const repos = new Array()
+    fetch(path)
+      .then((response) => {
+        if (response.ok) {
+          return response.json()
         } else {
-          this.setState({
-            displayLoadIcon: false,
-            isEmptyResults: false,
-            selectAllCheckValue: false,
-          })
+          throw new Error(response.statusText)
         }
       })
-      .catch(error => {
-        // might be a timeout error
+      .then(responseJSON => {
+        for (const repository of responseJSON.__children__) {
+          repos.push({ name: repository.__name__, id: repository["jcr:uuid"] })
+        }
         this.setState({
-          displayLoadIcon: false,
-          isSearchException: true
-        },()=>{ console.log("[doSearch] error ", error) })
-
+          repositories: repos,
+          filteredRepositories: repos
+        })
       })
+      .catch((error) => {
+        console.log(error)
+      })
+
   }
 
-  private dismissNotification = () => {
-    this.setState({ isEmptyResults: false, isSearchException: false });
+  private getProducts = () => {
+    const path = "/content/products.harray.1.json"
+    const products = new Array()
+    fetch(path)
+      .then((response) => {
+        if (response.ok) {
+          return response.json()
+        } else {
+          throw new Error(response.statusText)
+        }
+      })
+      .then(responseJSON => {
+        for (const product of responseJSON.__children__) {
+          products.push({ name: product.__name__, id: product["jcr:uuid"] })
+        }
+        this.setState({ products })
+      })
+      .catch((error) => {
+        console.log(error)
+      })
+
+  }
+  private onExpand = () => {
+    this.drawerRef.current && this.drawerRef.current.focus()
   };
 
-  private setQuery = (prod: string) => {
-    this.setState({ filterQuery: prod })
+  private onClick = () => {
+    const isExpanded = !this.state.isExpanded;
+    this.setState({
+      isExpanded
+    });
   };
 
-  private buildSearchUrl() {
-    let backend = "/pantheon/internal/modules.json?"
-    backend += this.state.filterQuery
-    if (this.state.filterQuery.trim() !== "") {
-      backend += "&"
-    }
-    backend += "offset=" + ((this.state.page - 1) * this.state.pageLimit) + "&limit=" + this.state.pageLimit
-    if (!backend.includes("Uploaded") && !backend.includes("direction")) {
-      backend += "&key=Uploaded&direction=desc"
-    }
+  private onCloseClick = () => {
+    this.setState({
+      isExpanded: false
+    });
+  };
 
-    return backend
+  // methods for toolbar
+  private onInputChange = newValue => {
+    this.setState({ inputValue: newValue });
+  };
+
+  private onInputClear = (event) => {
+    this.setState({ inputValue: "" })
   }
 
-  private updatePageCounter = (direction: string) => () => {
-    if (direction === "L" && this.state.page > 1) {
-      this.setState({ page: this.state.page - 1 }, () => {
-        this.doSearch()
-      })
-    } else if (direction === "R") {
-      this.setState({ page: this.state.page + 1 }, () => {
-        this.doSearch()
-      })
-    } else if (direction === "F") {
-      this.setState({ page: 1 }, () => {
-        this.doSearch()
-      })
+  private onSelect = (type, event, selection) => {
+    const checked = event.target.checked;
+    this.setState(prevState => {
+      const prevSelections = prevState.filters[type];
+      return {
+        filters: {
+          ...prevState.filters,
+          [type]: checked ? [...prevSelections, selection] : prevSelections.filter(value => value !== selection)
+        }
+      };
+    });
+  };
+
+  private onStatusSelect = (event, selection) => {
+    this.onSelect("status", event, selection);
+  };
+
+  private onCtypeSelect = (event, selection) => {
+    this.onSelect("ctype", event, selection);
+  };
+
+  private onDelete = (type: string | ToolbarChipGroup = "", id: string | ToolbarChip = "") => {
+    if (type) {
+      let filterType
+      filterType = typeof type === "object" ? type.name : type
+      filterType = type === 'Content Type' ? 'ctype' : type
+      this.setState(prevState => {
+        const newState = Object.assign(prevState);
+        newState.filters[filterType.toLowerCase()] = newState.filters[filterType.toLowerCase()].filter(s => s !== id);
+        return {
+          filters: newState.filters
+        };
+      });
+    } else {
+      this.setState({
+        filters: {
+          ctype: [],
+          status: []
+        }
+      });
     }
-  }
+  };
 
-  private changePerPageLimit = (pageLimitValue) => {
-    this.setState({ pageLimit: pageLimitValue, page: 1 }, () => {
-      // console.log("pageLImit value on calling changePerPageLimit function: "+this.state.pageLimit)
-      return (this.state.pageLimit + " items per page")
-    })
-  }
+  private onDeleteGroup = type => {
+    let filterType
+    filterType = type === 'Content Type' ? 'ctype' : type
+    this.setState(prevState => {
+      prevState.filters[filterType.toLowerCase()] = [];
+      return {
+        filters: prevState.filters
+      };
+    });
+  };
 
-  private buildTransientPathArray = () => {
-    const tPaths: string[] = []
-    this.state.results.map(item => {
-      if (item[Search.KEY_CHECKEDITEM]) {
-        tPaths.push(item[Search.KEY_TRANSIENTPATH])
+  private onStatusToggle = isExpanded => {
+    this.setState({
+      statusIsExpanded: isExpanded
+    });
+  };
+
+  private onCtypeToggle = isExpanded => {
+    this.setState({
+      ctypeIsExpanded: isExpanded
+    });
+  };
+
+  // methods for filter search
+  private onChangeRepoFilter = (value, event) => {
+    this.setState({
+      repoFilterValue: value
+    });
+
+    // check for input value
+    if (value) {
+      // filter and return repositories that include input value, and set state to the filtered list
+      let filtered = this.state.repositories.filter(data => data.name.toLowerCase().includes(value.toLowerCase()))
+      this.setState({
+        filteredRepositories: filtered
+      })
+    } else {
+      this.getRepositories()
+    }
+  };
+
+  private onChangeProductFilter = (value, event) => {
+    this.setState({
+      productFilterValue: value
+    });
+
+    if (value) {
+      let inputString = "";
+      const matchFound = [{ name: "", id: "" }];
+
+      this.state.products.map(data => {
+        inputString = "" + data.name
+        if (inputString.toLowerCase().includes(value.toLowerCase())) {
+          matchFound.push(data)
+        }
+      });
+      this.setState({ products: matchFound })
+    } else {
+      this.getProducts()
+    }
+  };
+
+  private onSelectRepositories = (checked, event) => {
+    let repositoriesSelected = new Array()
+    let repositories
+
+    repositories = this.state.repositories.map(item => {
+      if (item.id === event.target.id) {
+        item.checked = checked; 
       }
-    })
-    return tPaths
+      return item;
+    });
+
+    repositoriesSelected = repositories.map(item => {
+      if (item.checked !== undefined && item.checked === true) {
+        if (item.name !== undefined) {
+          return item.name
+        }
+      }
+    });
+
+    // filter undefined values
+    repositoriesSelected = repositoriesSelected.filter(r => r !== undefined)
+
+    this.setState({
+      repositories,
+      repositoriesSelected
+    });
+
   }
 
+  // Method for ExpandableSection
+  private onExpandableToggle = isExpanded => {
+    this.setState({
+      expandableSectionIsExpanded: isExpanded
+    });
+  };
+
+  private onModulesToggle = () => {
+    const modulesIsExpanded = !this.state.modulesIsExpanded
+    this.setState({
+      modulesIsExpanded
+    });
+  };
+
+  private onAssembliesToggle = () => {
+    const assembliesIsExpanded = !this.state.assembliesIsExpanded
+    this.setState({
+      assembliesIsExpanded
+    });
+  };
+
+  private onRepositoriesToggle = () => {
+    const repoFilterIsExpanded = !this.state.repoFilterIsExpanded
+    this.setState({
+      repoFilterIsExpanded
+    });
+  };
 }
 
-export { Search }
+
+export { Search };
